+++ DeltaEngine/DeltaGraphics/src/EngineCore/Rendering/OpenGL/GlBufferLayout.cpp
@@ -1,4 +1,3 @@
-#include "PCH.hpp"
 #include "EngineCore/Rendering/BufferLayout.hpp"
 
 #include <glad/glad.h>
@@ -15,7 +14,7 @@ void ShaderData::Construct(Type shaderDataType)
     case Type::INT3:
     case Type::INT4:
         m_Type = GL_INT;
-        m_Count = ((GLint)shaderDataType - (GLint)Type::INT + 1);
+        m_Count = ((size_t)shaderDataType - (size_t)Type::INT + 1);
         m_Size = sizeof(GLint) * m_Count;
         break;
     case Type::UINT:
@@ -23,7 +22,7 @@ void ShaderData::Construct(Type shaderDataType)
     case Type::UINT3:
     case Type::UINT4:
         m_Type = GL_UNSIGNED_INT;
-        m_Count = ((GLint)shaderDataType - (GLint)Type::UINT + 1);
+        m_Count = ((size_t)shaderDataType - (size_t)Type::UINT + 1);
         m_Size = sizeof(GLuint) * m_Count;
         break;
     case Type::FLOAT:
@@ -31,7 +30,7 @@ void ShaderData::Construct(Type shaderDataType)
     case Type::FLOAT3:
     case Type::FLOAT4:
         m_Type = GL_FLOAT;
-        m_Count = ((GLint)shaderDataType - (GLint)Type::FLOAT + 1);
+        m_Count = ((size_t)shaderDataType - (size_t)Type::FLOAT + 1);
         m_Size = sizeof(GLfloat) * m_Count;
         break;
     }
@@ -51,7 +50,7 @@ void BufferLayout::Init(const std::vector<ShaderData::Type>& shaderDataTypes)
 {
     m_Elements.reserve(shaderDataTypes.size());
 
-    GLint offset = 0;
+    size_t offset = 0;
     for (const auto& shaderDataType : shaderDataTypes)
     {
         BufferElement element;
